backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
